refactor(client): split SpacesComponent into per-concern components (A17.8b-2b)
Completes A17.8. spaces.component.ts: 1893 -> 262. The create dialog and the four
settings tabs are their own components; the page is now just the space list.

  space-schema-tab       795  (+ import-conflict and library-picker dialogs)
  space-create-dialog    225
  space-danger-tab       195
  space-duplicates-tab   144
  space-settings-tab      64
  space-dialog.styles    102  (shared const)
  spaces.component       262  (the list)

The services from the earlier parts paid off exactly as intended: because
SpacesStore owns the server data and SpaceSettingsState owns the dialog form
state, the children need NO data @Output() plumbing. The only output anywhere is
the create dialog's `closed` — its visibility really is the page's view state, so
that one belongs.

Styles: Angular scopes component styles, so each child rendering the chip inputs
or sp-* chrome needs the rules in its own metadata. Rather than paste ~90 lines
into six components — five copies free to drift — they share one
SPACE_DIALOG_STYLES const. The repo styles inline everywhere (no styleUrls), so a
const keeps the convention while staying DRY. Verified AOT resolves it before
building six components on the assumption.

OnPush: every extracted component is OnPush from birth and asserted
(space-components.onpush.spec.ts), matching brain/file-manager/graph/audit-log.
The old monolith was the one major page that was not, and could not be flipped as
it stood — FileReader.onload mutated schImportError/schImportInfo with no signal
write, so the import result would have silently stopped rendering. Extraction made
it tractable: those are signals now, in the schema tab where they live. The 262-line
shell is still default CD; flipping it is small and safe now and is tracked, not
bundled.

All 36 characterization assertions from #237 survive, each moved with the code it
covers, none rewritten to make a refactor pass: 14 on the page (list view state +
rendering), 22 on SpaceSettingsState, 2 on the create dialog.

Build warnings: fixed the two I introduced (unused PhIconComponent) and one
pre-existing NG8102 in pages/schema-library. The build is now warning-free except
`qrcode` shipping CommonJS (an optimization bailout) — logged as A20 with the real
fix (an ESM QR library) rather than silenced with allowedCommonJsDependencies.

Verified: client build clean; suite 121 -> 127, all green (18 files). Docs checked
— no doc references these internals.
